refactor(app): local api instead of global api
