--- src/api/index.tsx
@@ -1,2 +1,0 @@
-export * from './auth';
-export * from './register';
